requirements.txt
apache-airflow
kafka-python==2.0.2
requests
boto3
pyarrow
kafka